SFT-100 Refactoring AuxBMS Code
Due to difficulties when debugging issues, it was decided to refactor the code to achieve some clarity.

Separated each of the states and functions into it's own files. Created functions for commonly used logical evaluations. Create a folder to organize all the SetContactorTask related files.
