+++ EpsilonAuxBMS/Inc/SetContactor/States/Blocked.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "StateFunctionsShared.h"
+
+void blocked(SetContactorStateInfo* stateInfo);
