+++ EpsilonAuxBMS/Inc/SetContactor/States/ChargeEnableCheck.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "StateFunctionsShared.h"
+
+void chargeEnableCheck(SetContactorStateInfo* stateInfo);
