+++ EpsilonAuxBMS/Inc/SetContactor/States/CommonEnableCheck.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "StateFunctionsShared.h"
+
+void commonEnableCheck(SetContactorStateInfo* stateInfo);
