+++ EpsilonAuxBMS/Inc/SetContactor/States/DischargeEnableCheck.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "StateFunctionsShared.h"
+
+void dischargeEnableCheck(SetContactorStateInfo* stateInfo);
