+++ EpsilonAuxBMS/Inc/SetContactor/States/FirstCheck.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "StateFunctionsShared.h"
+
+void firstCheck(SetContactorStateInfo* stateInfo);
