--- EpsilonAuxBMS/Tests/tests/SetContactorsTaskTest/SetContactorsTaskTest.h
@@ -1,4 +1,0 @@
-#pragma once
-
-void runSetContactorsTests();
-void test_disconnectContactors();
